Update patch set 1
Patch Set 1:

It's part of core tutorials, so the artifact id must be coretutorials, for which the ODL nexus repo is created. 

the python script is slightly different from the one in integration; you ccould say this is a newer version of the integraion one and it works with the benchmark embedded in the controller. There could be just one, benhcmark script but it's more convenient for a novice user to have the script here.

Patch-set: 1