start all hs-sample projects
HighScalability is possible if we use NodeId as a GroupServiceKey
for all cooperate application. Every application has to have two
paths "Manager" class which is AppBase and has same lifeCycle as
the application. So every ClusterNode has own Manager instance.
And second parth "Context" which represent exactly one Device
and "Context" instance has to be a ClusterSingletonService.

Change-Id: Iff573c30eb3fb500244233e3bb9e594793c619f3
Signed-off-by: Vaclav Demcak <vdemcak@cisco.com>