Unify versions to 2.2.0 for Carbon-SR2-based released
We are adopting a new versioning scheme, where each upstream release
receives a major version bump to coretutorials. We also use unified
version for all components, so we have an easier time.

Since 1.0.x was Lithium-based and Beryllium and Boron are dead now,
use 2.2.0 for Carbon-SR2. 2.3.0 will be for Carbon-SR3, 3.0.0
for Nitrogen, 3.1.0 for Nitrogen-SR1, etc.

Change-Id: I08556a276e42cf13287e3ee5503bbbf926fcf491
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>