Bump to Oxygen GA MSI artifacts
Update towards Oxygen requires us to align with our immediate
upstreams. The first step is to point towards MSI Oxygen GA versions,
which provide the minimum compile surface.

This also disables feature packaging, as that has been reworked
in Oxygen and we are not ready fo that. We will re-introduce it once
the code actually builds.

Change-Id: Ie737a3b14ed65b7c424be65bfea4af5dc8b31f2f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>